I stand here humbled because I am now back in the transformation stage, which I thought I had left in the past.
After starting to design the layout of the NYCTaxiCab report, I discovered that I was overlooking a very valuable data point: the location data (i.e., start latitude and longitude and end latitude and longitude).

This led me to consolidate the four location columns in the fact table into two columns: start_location and end_location (previously start_lat, start_lon, end_lat, and end_lon). Next, I ingested borough and neighborhood data using the NYC Open Data website. This was done with the hope of being able to filter the fact data by location, specifically by borough and neighborhood. Latitude and longitude are not very useful in reports unless they are associated with some kind of nomenclature. Below is the code that allowed me to populate the start_location and end_location tables and add them to the database, ans also the bash code to the add the geo.json files to the postgresql database.